# services/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOME/IP Simulation Studio",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Projects",
        "operationId": "getProjects",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCollection"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/import": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Import Project",
        "operationId": "importProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/overview": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Overview",
        "operationId": "getProjectOverview",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/pipeline": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Project Pipeline",
        "operationId": "getProjectPipeline",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/pipeline/generate": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Generate Artifacts",
        "operationId": "generateProjectArtifacts",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/pipeline/build": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Build Nodes",
        "operationId": "buildProjectNodes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectBuildRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiError"
      },
...
      "DocumentRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "DocumentRef"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterfaceIndex": {
        "properties": {
          "interfaces": {
            "items": {
              "$ref": "#/components/schemas/InterfaceSummary"
            },
            "type": "array",
            "title": "Interfaces"
          },
          "deployments": {
            "items": {
              "$ref": "#/components/schemas/DeploymentSummary"
            },
            "type": "array",
            "title": "Deployments"
          }
        },
        "type": "object",
        "title": "InterfaceIndex"
      },
...
      "Job": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "log": {
            "type": "string",
            "title": "Log",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "type",
          "status"
        ],
        "title": "Job"
      },
      "NetworkSettings": {
        "properties": {
          "service_discovery": {
            "type": "boolean",
            "title": "Service Discovery",
            "default": true
          },
          "multicast": {
            "type": "string",
            "title": "Multicast",
            "default": "224.224.224.245"
          },
          "sd_port": {
            "type": "integer",
            "title": "Sd Port",
            "default": 30490
          }
        },
        "type": "object",
        "title": "NetworkSettings"
      },
...
      "PipelineOverview": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "status"
        ],
        "title": "PipelineOverview"
      },
...
      "ProjectBuildRequest": {
        "properties": {
          "clean": {
            "type": "boolean",
            "title": "Clean",
            "default": false
          },
          "generate": {
            "type": "boolean",
            "title": "Generate",
            "default": true
          }
        },
        "type": "object",
        "title": "ProjectBuildRequest"
      },
      "ProjectCollection": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "type": "array",
            "title": "Projects"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/StarterPreset"
            },
            "type": "array",
            "title": "Presets"
          }
        },
        "type": "object",
        "title": "ProjectCollection"
      },
      "ProjectCreateRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name"
        ],
        "title": "ProjectCreateRequest"
      },
      "ProjectImportRequest": {
        "properties": {
          "source_path": {
            "type": "string",
            "title": "Source Path"
          }
        },
        "type": "object",
        "required": [
          "source_path"
        ],
        "title": "ProjectImportRequest"
      },
      "ProjectOverview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "readiness": {
            "type": "integer",
            "title": "Readiness"
          },
          "readiness_gates": {
            "items": {
              "$ref": "#/components/schemas/ReadinessGate"
            },
            "type": "array",
            "title": "Readiness Gates"
          },
          "recommended_action": {
            "$ref": "#/components/schemas/RecommendedAction"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkSettings"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "interface_index": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterfaceIndex"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRef"
            },
            "type": "array",
            "title": "Documents"
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count",
            "default": 0
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Recent Runs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "readiness",
          "recommended_action",
          "network"
        ],
        "title": "ProjectOverview"
      },
      "ProjectStatus": {
        "properties": {
          "validation": {
            "type": "string",
            "title": "Validation"
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "build": {
            "type": "string",
            "title": "Build"
          },
          "last_run": {
            "type": "string",
            "title": "Last Run"
          }
        },
        "type": "object",
        "required": [
          "validation",
          "generation",
          "build",
          "last_run"
        ],
        "title": "ProjectStatus"
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count",
            "default": 0
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count",
            "default": 0
          },
          "latest_status": {
            "$ref": "#/components/schemas/ProjectStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "latest_status"
        ],
        "title": "ProjectSummary"
      },
      "ReadinessGate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "status",
          "ready",
          "detail"
        ],
        "title": "ReadinessGate"
      },
      "RecommendedAction": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "author",
              "build",
              "simulate",
              "inspect"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "title",
          "detail"
        ],
        "title": "RecommendedAction"
      },
...
      "StarterPreset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "sample"
          },
          "runnable": {
            "type": "boolean",
            "title": "Runnable",
            "default": false
          },
          "default_project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Project Id"
          },
          "default_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StarterPreset"
      },
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status",
          "message"
        ],
        "title": "TaskResponse"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "TaskStatus"
      },
      "TopologyNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "interface": {
            "type": "string",
            "title": "Interface"
          },
          "app_name": {
            "type": "string",
            "title": "App Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "interface",
          "app_name"
        ],
        "title": "TopologyNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}